moved formats into constant file and get_time_function in utils file
