+++ pyQuARC/code/constants.py
@@ -77,3 +77,13 @@ GCMD_LINKS = {
 }
 
 CMR_URL = "https://cmr.earthdata.nasa.gov"
+
+DATE_FORMATS = [
+    "%Y-%m-%dT%H:%M:%S.%f",  # Year to microsecond
+    "%Y-%m-%dT%H:%M:%S",  # Year to second
+    "%Y-%m-%dT%H:%M",  # Year to minute
+    "%Y-%m-%dT%H",  # Year to hour
+    "%Y-%m-%d",  # Year to day
+    "%Y-%m",  # Year to month
+    "%Y",  # Year
+]
